+++ RELEASE_NOTES.md
@@ -1,5 +1,11 @@
 #
 
+## Release Notes - Version 1.0.0
+
+- Function ```colorate``` is now fully functional and can format any string in any color for background and foreground.[^Note]
+
+[^Note]: The output color might differ from the desired one depending on terminal used.
+
 ## Release Notes - Version 0.0.2
 
 - Color list is now stored in a json file